+++ docs/changelog.rst
@@ -18,6 +18,7 @@ New features
 * Parameter ``threshold`` for determining the levels of a contour plot in class :class:`aspecd.plotting.SinglePlotter2D`
 * Parameter ``units`` in :class:`aspecd.annotation.Text`, :class:`aspecd.annotation.Marker`, :class:`aspecd.annotation.VerticalLine`, :class:`aspecd.annotation.HorizontalLine`, :class:`aspecd.annotation.VerticalSpan`, and  :class:`aspecd.annotation.HorizontalSpan` allowing to position using axes rather than data coordinates.
 * :class:`aspecd.processing.SliceRearrangement` for rearranging slices of a dataset along one dimension.
+* :class:`aspecd.processing.DatasetAlgebra` operates on a list of datasets, allowing to add/subtract multiple datasets from a given dataset.
 
 
 Fixes